refactor: move S3Uri to utils
